requirements.txt
fastapi
uvicorn[standard]
dnspython>=2.4.2
aiohttp>=3.9.0
pyyaml>=6.0.1
async-timeout>=4.0.3
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.4
transformers>=4.31.0
pillow>=10.0.0
qrcode>=7.4.2
numpy>=1.24.0
aiohttp>=3.8.5
pydantic>=2.1.1
accelerate>=0.21.0
qiskit>=0.44.0
qiskit-aer>=0.12.0
scipy>=1.11.0